Some Tested
### Fixed Issues
- The db id is included in login req/res now.
- The new accout and user are created normally now.

### Tested List
- Login: Login and Creating new account.
- RoomList: The server sends the room list that the user is in. The client shows the list.
- Create Room: Requesting creating room works well now.
- Enter Room: Requesting entering room works well now.
- Chat: Client can send a message in the room and receives the response of the sent message. (If the message is sent successfully, the loading icon will be turned check icon.)
- All jobs are stored in database.

+++ ChatClient/Assets/Scripts/Data/Room.cs
@@ -12,7 +12,7 @@ public class Room
 
     public ICollection<UserInfo> Users => Info.Users;
     public ulong RoomDbId => Info.RoomDbId;
-    public ulong RoomNumber => Info.RoomNumber;
+    public uint RoomNumber => Info.RoomNumber;
     public string RoomName => Info.RoomName;
 
     public bool Activated => ui != null && ui.gameObject.activeSelf;
